Delete the Middle Node of a Linked List - Day 10
What I need to do:
I am given the head of a singly linked list. My task is to delete the middle node of this linked list and return the head of the updated list.

The "middle node" is defined as the node at index floor(n / 2), where n is the total number of nodes in the list and indexing starts from 0. The floor function means we take the greatest integer less than or equal to n / 2.

For example:

If n = 1, the middle is at index 0.

If n = 2, the middle is at index 1.

If n = 3, the middle is at index 1.

If n = 4, the middle is at index 2.

If n = 5, the middle is at index 2.

If the list has only one node, then we delete that node and return nullptr.

To solve this, I will use the slow and fast pointer method to locate the middle node in one traversal, and then remove it by adjusting the previous node’s pointer.